+++ includes/wizards/class-reader-revenue-wizard.php
@@ -193,7 +193,7 @@ class Reader_Revenue_Wizard extends Wizard {
 			]
 		);
 
-		// Save a subscription.
+		// Update Donations settings.
 		register_rest_route(
 			NEWSPACK_API_NAMESPACE,
 			'/wizard/' . $this->slug . '/donations/',